// src/api-client/models/ThermostatType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const ThermostatType = {
  OpenthermRoomTemperature: "opentherm_room_temperature",
  OpenthermWithoutRoomTemperature: "opentherm_without_room_temperature",
  Null: "null",
} as const;
export type ThermostatType =
  (typeof ThermostatType)[keyof typeof ThermostatType];

export function ThermostatTypeFromJSON(json: any): ThermostatType {
  return ThermostatTypeFromJSONTyped(json, false);
}

export function ThermostatTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ThermostatType {
  return json as ThermostatType;
}

export function ThermostatTypeToJSON(value?: ThermostatType | null): any {
  return value as any;
}